# openapi-docs/edgen-api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Edgen API",
    "description": "Edgen API with OpenAI-compliant and proprietary endpoints.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/audio/transcriptions": {
      "post": {
        "tags": [
          "audio"
        ],
        "summary": "POST `/v1/audio/transcriptions`: transcribes audio into text.",
        "description": "POST `/v1/audio/transcriptions`: transcribes audio into text.\n\nSee [the original OpenAI API specification][openai], which this endpoint is compatible with.\n\n[openai]: https://platform.openai.com/docs/api-reference/auddio/createTranscription\n\nOn failure, may raise a `500 Internal Server Error` with a JSON-encoded [`WhisperEndpointError`]\nto the peer.",
        "operationId": "create_transcription",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTranscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "unexpected internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhisperEndpointError"
                }
              }
            }
          }
        }
      }
    },
    "/chat/completions": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "POST `/v1/chat/completions`: generate chat completions for the provided context, optionally",
        "description": "POST `/v1/chat/completions`: generate chat completions for the provided context, optionally\nstreaming those completions in real-time.\n\nSee [the original OpenAI API specification][openai], which this endpoint is compatible with.\n\n[openai]: https://platform.openai.com/docs/api-reference/chat/create\n\nGenerates completions for the given [`CreateChatCompletionRequest`] body.\nIf `stream` is enabled, streams a number of newline-separated, JSON-encoded\n[`ChatCompletionChunk`]s to the client using [server-sent events][sse]. Otherwise, returns a\nsingle JSON-encoded [`ChatCompletion`].\n\n[sse]: https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events\n\nOn failure, may raise a `500 Internal Server Error` with a JSON-encoded [`ChatCompletionError`]\nto the peer.",
        "operationId": "chat_completions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletion"
                }
              }
            }
          },
          "500": {
            "description": "unexpected internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Parse"
            ],
            "properties": {
              "Parse": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "ResamplerInit"
            ],
            "properties": {
              "ResamplerInit": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Resample"
            ],
            "properties": {
              "Resample": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ChatCompletion": {
        "type": "object",
        "description": "A fully generated chat completion.",
        "required": [
          "id",
          "choices",
          "created",
          "model",
          "system_fingerprint",
          "object",
          "usage"
        ],
        "properties": {
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatCompletionChoice"
            },
            "description": "The tokens generated by the model."
          },
          "created": {
            "type": "integer",
            "format": "int64",
            "description": "The UNIX timestamp at which the completion was generated."
          },
          "id": {
            "type": "string",
            "description": "A unique identifier for this completion."
          },
          "model": {
            "type": "string",
            "description": "The model that generated the completion."
          },
          "object": {
            "type": "string",
            "description": "The object type. This is always `text_completion`."
          },
          "system_fingerprint": {
            "type": "string",
            "description": "A unique identifier for the backend configuration that generated the completion."
          },
          "usage": {
            "$ref": "#/components/schemas/ChatCompletionUsage"
          }
        }
      },
...
      "ChatCompletionError": {
        "oneOf": [
          {
            "type": "object",
            "description": "The provided model could not be found on the local system.",
            "required": [
              "model_name",
              "error"
            ],
            "properties": {
              "error": {
                "type": "string",
                "enum": [
                  "no_such_model"
                ]
              },
              "model_name": {
                "type": "string",
                "description": "The name of the model."
              }
            }
          },
          {
            "type": "object",
            "description": "The provided model name contains prohibited characters.",
            "required": [
              "model_name",
              "reason",
              "error"
            ],
            "properties": {
              "error": {
                "type": "string",
                "enum": [
                  "prohibited_name"
                ]
              },
              "model_name": {
                "type": "string",
                "description": "The name of the model provided."
              },
              "reason": {
                "type": "string",
                "description": "A human-readable error message."
              }
            }
          },
          {
            "type": "object",
            "required": [
              "error"
            ],
            "properties": {
              "error": {
                "type": "string",
                "enum": [
                  "ffi"
                ]
              }
            }
          }
        ],
        "description": "An error condition raised by the chat completion API.\n\nThis is **not normative** with OpenAI's specification, which does not document any specific\nfailure modes.",
        "discriminator": {
          "propertyName": "error"
        }
      },
...
      "ChatMessages": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ChatMessage"
        },
        "description": "A sequence of chat messages in a [`CreateChatCompletionRequest`].\n\nThis implements [`Display`] to generate a transcript of the chat messages compatible with most\nLLaMa-based models."
      },
...
      "CreateChatCompletionRequest": {
        "type": "object",
        "description": "A request to generate chat completions for the provided context.\n\nAn `axum` handler, [`chat_completions`][chat_completions], is provided to handle this request.\n\nSee [the documentation for creating chat completions][openai] for more details.\n\n[chat_completions]: fn.chat_completions.html\n[openai]: https://platform.openai.com/docs/api-reference/chat/create",
        "required": [
          "model"
        ],
        "properties": {
          "frequency_penalty": {
            "type": "number",
            "format": "float",
            "description": "A number in `[-2.0, 2.0]`. A higher number decreases the likelihood that the model\nrepeats itself.",
            "nullable": true
          },
          "logit_bias": {
            "type": "object",
            "description": "A map of token IDs to `[-100.0, +100.0]`. Adds a percentage bias to those tokens before\nsampling; a value of `-100.0` prevents the token from being selected at all.\n\nYou could use this to, for example, prevent the model from emitting profanity.",
            "additionalProperties": {
              "type": "number",
              "format": "float"
            },
            "nullable": true
          },
          "max_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The maximum number of tokens to generate. If `None`, terminates at the first stop token\nor the end of sentence.",
            "nullable": true,
            "minimum": 0
          },
          "messages": {
            "$ref": "#/components/schemas/ChatMessages"
          },
          "model": {
            "type": "string",
            "description": "The model to use for generating completions."
          },
          "n": {
            "type": "number",
            "format": "float",
            "description": "How many choices to generate for each token in the output. `1` by default. You can use\nthis to generate several sets of completions for the same prompt.",
            "nullable": true
          },
          "presence_penalty": {
            "type": "number",
            "format": "float",
            "description": "A number in `[-2.0, 2.0]`. Positive values \"increase the model's likelihood to talk about\nnew topics.\"",
            "nullable": true
          },
          "response_format": {
            "description": "The format of the response stream.\n\nThis is always assumed to be JSON, which is non-conformant with the OpenAI spec.",
            "nullable": true
          },
          "seed": {
            "type": "integer",
            "format": "int32",
            "description": "An RNG seed for the session. Random by default.",
            "nullable": true,
            "minimum": 0
          },
          "stop": {
            "type": "string",
            "description": "A stop phrase or set of stop phrases.\n\nThe server will pause emitting completions if it appears to be generating a stop phrase,\nand will terminate completions if a full stop phrase is detected.\n\nStop phrases are never emitted to the client."
          },
          "stream": {
            "type": "boolean",
            "description": "If `true`, emit [`ChatCompletionChunk`]s instead of a single [`ChatCompletion`].\n\nYou can use this to live-stream completions to a client.",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "float",
            "description": "The sampling temperature, in `[0.0, 2.0]`. Higher values make the output more random.",
            "nullable": true
          },
          "tool_choice": {
            "type": "string",
            "description": "If present, the tool that the user has chosen to use.\n\nOpenAI states:\n\n- `none` prevents any tool from being used,\n- `auto` allows any tool to be used, or\n- you can provide a description of the tool entirely instead of a name."
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolStub"
            },
            "description": "A list of tools made available to the model.",
            "nullable": true
          },
          "top_p": {
            "type": "number",
            "format": "float",
            "description": "Nucleus sampling. If you set this value to 10%, only the top 10% of tokens are used for\nsampling, preventing sampling of very low-probability tokens.",
            "nullable": true
          },
          "user": {
            "type": "string",
            "description": "A unique identifier for the _end user_ creating this request. This is used for telemetry\nand user tracking, and is unused within Edgen.",
            "nullable": true
          }
        }
      },
      "CreateTranscriptionRequest": {
        "type": "object",
        "description": "A request to transcribe an audio file into text in either the specified language, or whichever\nlanguage is automatically detected, if none is specified.\n\nAn `axum` handler, [`create_transcription`][create_transcription], is provided to handle this request.\n\nSee [the documentation for creating transcriptions][openai] for more details.\n\n[create_transcription]: fn.create_transcription.html\n[openai]: https://platform.openai.com/docs/api-reference/audio/createTranscription",
        "required": [
          "file",
          "model"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "The audio file object (not file name) to transcribe, in one of these formats: `aac`, `flac`,\n`mp3`, `m4a`, `m4b`, `ogg`, `oga`, `mogg`, `wav`, `webm` TODO check working formats."
          },
          "language": {
            "type": "string",
            "description": "The language of the input audio. Supplying the input language in ISO-639-1 format will\nimprove accuracy and latency.",
            "nullable": true
          },
          "model": {
            "type": "string",
            "description": "ID of the model to use."
          },
          "prompt": {
            "type": "string",
            "description": "An optional text to guide the model's style or continue a previous audio segment. The prompt\nshould match the audio language.",
            "nullable": true
          },
          "response_format": {
            "type": "string",
            "description": "The format of the transcript output, in one of these options: json, text, srt, verbose_json,\nor vtt. TODO whats this?",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "float",
            "description": "The sampling temperature, between 0 and 1. Higher values like 0.8 will make the output more\nrandom, while lower values like 0.2 will make it more focused and deterministic. If set to 0,\nthe model will use log probability to automatically increase the temperature until certain\nthresholds are hit.",
            "nullable": true
          }
        }
      },
      "DecodeSessionError": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "SessionOver"
            ]
          },
          {
            "type": "object",
            "required": [
              "Send"
            ],
            "properties": {
              "Send": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "SessionRunner"
            ],
            "properties": {
              "SessionRunner": {
                "$ref": "#/components/schemas/SessionRunnerError"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Join"
            ],
            "properties": {
              "Join": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "ModelError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "FileNotFound"
            ],
            "properties": {
              "FileNotFound": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "UnknownModel"
            ],
            "properties": {
              "UnknownModel": {
                "$ref": "#/components/schemas/ModelKind"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "API"
            ],
            "properties": {
              "API": {
                "type": "string"
              }
            }
          },
          {
            "type": "string",
            "enum": [
              "NotPreloaded"
            ]
          }
        ]
      },
...
      "SessionRunnerError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Executor"
            ],
            "properties": {
              "Executor": {
                "$ref": "#/components/schemas/WhisperError"
              }
            }
          }
        ]
      },
...
      "WhisperEndpointError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "FileNotFound"
            ],
            "properties": {
              "FileNotFound": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "SessionNotFound"
            ],
            "properties": {
              "SessionNotFound": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Internal"
            ],
            "properties": {
              "Internal": {
                "$ref": "#/components/schemas/WhisperError"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Session"
            ],
            "properties": {
              "Session": {
                "$ref": "#/components/schemas/DecodeSessionError"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Audio"
            ],
            "properties": {
              "Audio": {
                "$ref": "#/components/schemas/AudioError"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Model"
            ],
            "properties": {
              "Model": {
                "$ref": "#/components/schemas/ModelError"
              }
            }
          }
        ]
      },
      "WhisperError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "UnsupportedMime"
            ],
            "properties": {
              "UnsupportedMime": {
                "type": "object",
                "required": [
                  "mime",
                  "supported"
                ],
                "properties": {
                  "mime": {
                    "type": "string"
                  },
                  "supported": {
                    "type": "string"
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "UnsupportedSampleRate"
            ],
            "properties": {
              "UnsupportedSampleRate": {
                "type": "object",
                "required": [
                  "value",
                  "supported"
                ],
                "properties": {
                  "supported": {
                    "type": "string"
                  },
                  "value": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Parsing"
            ],
            "properties": {
              "Parsing": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Internal"
            ],
            "properties": {
              "Internal": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "ModelInitialization"
            ],
            "properties": {
              "ModelInitialization": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "SessionInitialization"
            ],
            "properties": {
              "SessionInitialization": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "Other"
            ],
            "properties": {
              "Other": {
                "type": "string"
              }
            }
          }
        ]
      },
...
    }
  }
}